assets/images/posts/2026/agentic-ai-system/agentcore-deploy/requirements.txt
strands-agents>=1.0.0
bedrock-agentcore
